requirements.txt
asgiref==3.8.1
autopep8==2.3.1
dj-database-url==2.3.0
Django==4.2.16
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-jazzmin==3.0.1
django-silk==5.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
et_xmlfile==2.0.0
gprof2dot==2024.6.6
gunicorn==23.0.0
inflection==0.5.1
numpy==2.0.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.0.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.1
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
whitenoise==6.8.2
